Add support for SLY
This change introduces a new customizable variable, `portacle-ide`,
that lets the user choose the IDE started when Portacle launches. The
default value is a function that prompts the user to choose between
SLIME or SLY and then asks the user if he'd like to keep that choice
for future sessions. SLIME is the first and default choice.

In any case, even if the user chooses SLIME, SLY is loaded alongside
it. This is quite fine as long as the `lisp-mode-hook` doesn't contain
both `slime-major-mode-hook` and `sly-editing-mode` simultaneously,
which would cause keybinding conflicts in Lisp source buffers. To
avoid this, very light advising of `sly` and `slime` functions is
used.

The variable `portacle-ide` is customizable, should the user change
her/his mind later on. However, to enable saving of customization
variables to actually work in Portacle's Emacs, which is launched with
custom-file but without user-init-file, we set the latter to the
former. This change is in portacle-user.el.

* portacle-slime.el (slime): Advise slime to watch out for SLY
major-mode hook.
(portacle-slime-contribs): Setup contribs when loading init.

* portacle-sly.el: New file. Same as portacle-slime, but simpler.

* portacle-user.el (file-exists-p): Set user-init-file to
something non-nil so that saving customization actually works.

* portacle.el (portacle-sly): Require it.
(portacle-let-user-choose-ide): New function.
(defgroup portacle): New customization group.
(defcustom portacle-ide): New custom variable.
(portacle--start-ide-maybe): New helper.
(emacs-startup-hook): Add it here.